+++ ms-inventario/src/main/java/cl/hilton/inventario/repository/ProductoRepository.java
@@ -1,6 +1,5 @@
 package cl.hilton.inventario.repository;
 
-
 import cl.hilton.inventario.model.Producto;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
@@ -9,7 +8,7 @@ import java.util.List;
 import java.util.Optional;
 
 @Repository
-public interface ProductoRepository extends JpaRepository<Producto, Integer> {
+public interface ProductoRepository extends JpaRepository<Producto, Long> {
 
     Optional<Producto> findByCodigoProducto(String codigoProducto);
 
@@ -19,11 +18,11 @@ public interface ProductoRepository extends JpaRepository<Producto, Integer> {
 
     List<Producto> findByUnidad(String unidad);
 
-    List<Producto> findByStockActualLessThanEqual(Integer stockActual);
+    List<Producto> findByStockActualLessThanEqual(Long stockActual);
 
-    List<Producto> findByStockActualLessThan(Integer stockActual);
+    List<Producto> findByStockActualLessThan(Long stockActual);
 
-    List<Producto> findByStockActualGreaterThan(Integer stockActual);
+    List<Producto> findByStockActualGreaterThan(Long stockActual);
 
     List<Producto> findByNombreContainingIgnoreCase(String nombre);
 }
