+++ ms-restaurante/src/main/java/cl/hilton/restaurante/repository/ItemPedidoRepository.java
@@ -1,7 +1,5 @@
 package cl.hilton.restaurante.repository;
 
-
-
 import cl.hilton.restaurante.model.ItemPedido;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
@@ -9,12 +7,11 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface ItemPedidoRepository extends JpaRepository<ItemPedido, Integer> {
+public interface ItemPedidoRepository extends JpaRepository<ItemPedido, Long> {
 
     List<ItemPedido> findByPedidoNumeroPedido(String numeroPedido);
 
     List<ItemPedido> findByNombreProductoContainingIgnoreCase(String nombreProducto);
 
-    List<ItemPedido> findByCantidadGreaterThan(Short cantidad);
+    List<ItemPedido> findByCantidadGreaterThan(Integer cantidad);
 }
-
